arch: extract JSON-RPC 2.0 types into dedicated module (#281)
Move generic JSON-RPC protocol types (Request, Response, Error,
MessageKind, classify/parse helpers) from acp.rs into app::jsonrpc
so they can be reused by future protocol implementations.
